assignments/class_08/pyproject.toml
[project]
name = "class-08"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.2.5",
    "pydantic>=2.11.7",
]
